Rename nonelab to fontshell (name idea from Antonio Cavedoni (@verbosus))
