+++ Lib/fontParts/fontshell/__init__.py
@@ -1,0 +1,17 @@
+from fontParts.base.errors import FontPartsError
+from fontParts.fontshell.font import RFont
+from fontParts.fontshell.info import RInfo
+from fontParts.fontshell.groups import RGroups
+from fontParts.fontshell.kerning import RKerning
+from fontParts.fontshell.features import RFeatures
+from fontParts.fontshell.lib import RLib
+from fontParts.fontshell.layer import RLayer
+from fontParts.fontshell.glyph import RGlyph
+from fontParts.fontshell.contour import RContour
+from fontParts.fontshell.point import RPoint
+from fontParts.fontshell.segment import RSegment
+from fontParts.fontshell.bPoint import RBPoint
+from fontParts.fontshell.component import RComponent
+from fontParts.fontshell.anchor import RAnchor
+from fontParts.fontshell.guideline import RGuideline
+from fontParts.fontshell.image import RImage
