+++ Lib/fontParts/fontshell/anchor.py
@@ -1,6 +1,6 @@
 import defcon
 from fontParts.base import BaseAnchor
-from fontParts.nonelab.base import RBaseObject
+from fontParts.fontshell.base import RBaseObject
 
 
 class RAnchor(RBaseObject, BaseAnchor):
