+++ Lib/fontParts/fontshell/component.py
@@ -1,6 +1,6 @@
 import defcon
 from fontParts.base import BaseComponent, FontPartsError
-from fontParts.nonelab.base import RBaseObject
+from fontParts.fontshell.base import RBaseObject
 
 
 class RComponent(RBaseObject, BaseComponent):
