+++ Lib/fontParts/fontshell/features.py
@@ -1,6 +1,6 @@
 import defcon
 from fontParts.base import BaseFeatures, FontPartsError
-from fontParts.nonelab.base import RBaseObject
+from fontParts.fontshell.base import RBaseObject
 
 
 class RFeatures(RBaseObject, BaseFeatures):
