+++ Lib/fontParts/fontshell/info.py
@@ -1,6 +1,6 @@
 import defcon
 from fontParts.base import BaseInfo, FontPartsError
-from fontParts.nonelab.base import RBaseObject
+from fontParts.fontshell.base import RBaseObject
 
 
 class RInfo(RBaseObject, BaseInfo):
