+++ Lib/fontParts/fontshell/segment.py
@@ -1,6 +1,6 @@
 import defcon
 from fontParts.base import BaseSegment, FontPartsError
-from fontParts.nonelab.base import RBaseObject
+from fontParts.fontshell.base import RBaseObject
 
 
 class RSegment(BaseSegment, RBaseObject): pass
