--- Lib/fontParts/nonelab/__init__.py
@@ -1,17 +1,0 @@
-from fontParts.base.errors import FontPartsError
-from fontParts.nonelab.font import RFont
-from fontParts.nonelab.info import RInfo
-from fontParts.nonelab.groups import RGroups
-from fontParts.nonelab.kerning import RKerning
-from fontParts.nonelab.features import RFeatures
-from fontParts.nonelab.lib import RLib
-from fontParts.nonelab.layer import RLayer
-from fontParts.nonelab.glyph import RGlyph
-from fontParts.nonelab.contour import RContour
-from fontParts.nonelab.point import RPoint
-from fontParts.nonelab.segment import RSegment
-from fontParts.nonelab.bPoint import RBPoint
-from fontParts.nonelab.component import RComponent
-from fontParts.nonelab.anchor import RAnchor
-from fontParts.nonelab.guideline import RGuideline
-from fontParts.nonelab.image import RImage
