+++ tox.ini
@@ -18,5 +18,5 @@ deps =
 install_command =
     pip install -v {opts} {packages}
 commands =
-    cov: coverage run Lib/fontParts/nonelab/test.py {posargs}
-    nocov: python Lib/fontParts/nonelab/test.py {posargs}
+    cov: coverage run Lib/fontParts/fontshell/test.py {posargs}
+    nocov: python Lib/fontParts/fontshell/test.py {posargs}
